app/requirements.txt
Flask==3.1.3
Flask-Session==0.8.0
Flask-WTF==1.2.2
supabase==2.3.4
requests==2.32.5
python-dotenv==1.0.1